Made skeleton code align with current file structure
